+++ funpayparsers/parsers/page_parsers/__init__.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from .chat_page_parser import *
 from .main_page_parser import *
 from .order_page_parser import *
-from .my_lots_page_parser import *
 from .profile_page_parser import *
 from .my_chips_page_parser import *
 from .settings_page_parser import *
+from .my_offers_page_parser import *
 from .subcategory_page_parser import *
 from .transactions_page_parser import *
