+++ funpayparsers/types/pages/__init__.py
@@ -4,9 +4,9 @@ from .base import *
 from .chat_page import *
 from .main_page import *
 from .order_page import *
-from .my_lots_page import *
 from .profile_page import *
 from .my_chips_page import *
 from .settings_page import *
+from .my_offers_page import *
 from .subcategory_page import *
 from .transactions_page import *
